Rename mobile components folder
